pokemon details template and style  - not done with the implementation
