--- pokemon-minisite/src/app/components/pokemon-detail/pokemon-detail.component.html
@@ -1,1 +1,0 @@
-<p>pokemon-detail works!</p>
