--- pokemon-minisite/src/app/components/pokemon-detail/pokemon-detail.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-pokemon-detail',
-  standalone: true,
-  imports: [],
-  templateUrl: './pokemon-detail.component.html',
-  styleUrl: './pokemon-detail.component.scss'
-})
-export class PokemonDetailComponent {
-
-}
