+++ pokemon-minisite/src/app/components/pokemon-list/pokemon-detail/pokemon-detail.component.scss
@@ -1,0 +1,68 @@
+@import "../../../styles.scss";
+
+.pokemon-popup-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: var(--popup-backgroundcolor);
+  @include centered-style;
+  z-index: 999;
+
+  .pokemon-popup {
+    background-color: var(--yellow-color);
+    border: none;
+    transition: background-color 0.3s ease;
+    width: 40%;
+    height: 40%;
+    border-radius: 10px;
+    box-shadow: var-(-box-shadow);
+  }
+
+  .pokemon-title {
+    font-family: var(--main-font);
+    background-color: var(--red-color);
+    border-radius: 10px;
+    font-weight: bold;
+    text-align: center;
+    font-size: 22px;
+    padding: 15px;
+    color: var(--white-color);
+  }
+
+  .card-body {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: center;
+    margin-top: 20px;
+  }
+
+  .pokemon-image {
+    width: 150px;
+    height: 150px;
+  }
+
+  .pokemon-details {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+  }
+
+  .pokemon-info {
+    display: flex;
+    margin-bottom: 10px;
+  }
+
+  .info-label {
+    font-weight: bold;
+    font-family: var(--main-font);
+  }
+
+  .info-value {
+    width: auto;
+    margin-left: 2px;
+    font-family: var(--main-font);
+  }
+}
